Make "day-of-month" behavior more consistent (#1)
Previously, the "day-of-month" behavior was broken at the end of the
month during long months before short months -- e.g. on January 30th if
you asked for Febuary 1st, it would return January 30th.

This makes it more consistent by setting the day of month to 1, then
adding a month, then adding (requested date - 1) days.

This is still broken if you ask for e.g. February 31st, but that is out
of scope of this PR.